nodes/VoiceStreamAI/requirements.txt
websockets
speechbrain
pyannote-audio
asyncio
sentence-transformers
transformers
faster-whisper
